# docs/openapi/prompts.yaml
openapi: 3.1.0
info:
  title: Prompts & Rendering API
  version: 1.0.0
paths:
...
  /v1/prompts/{id}/versions:
    post:
      summary: Create a Prompt Version
      description: Creates a new version of the prompt template in 'draft' state.
      operationId: createVersion
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique prompt UUID.
          schema:
            type: string
            format: uuid
      x-edge-cases:
        - "Validates template string syntax using Go's text/template parser before creating the entry."
      x-test-coverage:
        - "TestCreateVersion_Success: Verifies version number is incremented to draft status 1."
        - "TestCreateVersion_InvalidTemplate: Verifies syntax check fails with 400 and 'invalid template'."
        - "TestCreateVersion_MissingTemplate: Verifies missing template rejects with 400 and 'template is required'."
        - "TestCreateVersion_PromptNotFound: Verifies 404 response on missing parent prompt UUID."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateVersionRequest"
      responses:
        "201":
          description: Version created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    $ref: "#/components/schemas/PromptVersion"
                required:
                  - version
        "400":
          description: Invalid template or syntax error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              examples:
                missing_template:
                  value: { "error": "template is required" }
                syntax_error:
                  value: { "error": "invalid template: template parse error..." }
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "prompt not found"
        "500":
          description: Internal error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "internal error"

    get:
      summary: List Prompt Versions
      description: Lists all template versions associated with a prompt container.
      operationId: listVersions
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique prompt UUID.
          schema:
            type: string
            format: uuid
      x-test-coverage:
        - "TestListVersions: Verifies active versions listing."
      responses:
        "200":
          description: List of versions
          content:
            application/json:
              schema:
                type: object
                properties:
                  versions:
                    type: array
                    items:
                      $ref: "#/components/schemas/PromptVersion"
                required:
                  - versions
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Prompt not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "prompt not found"
  /v1/prompts/{id}/versions/{version}:
    get:
      summary: Get Prompt Version
      description: Retrieves details of a specific prompt template version by version number.
      operationId: getVersion
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique prompt UUID.
          schema:
            type: string
            format: uuid
        - name: version
          in: path
          required: true
          description: Version sequence number.
          schema:
            type: integer
      x-test-coverage:
        - "TestGetVersion: Verifies details retrieval of specific version."
        - "TestGetVersion_NotFound: Verifies 404 response on missing version sequence number."
      responses:
        "200":
          description: Prompt version details
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    $ref: "#/components/schemas/PromptVersion"
                required:
                  - version
        "400":
          description: Invalid path formatting
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "invalid version number"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Version not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "version not found"

    put:
      summary: Update Prompt Version Draft
      description: |
        Updates the draft template code of a specific version.
        Only versions currently in the 'draft' status may be modified.
      operationId: updateVersion
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique prompt UUID.
          schema:
            type: string
            format: uuid
        - name: version
          in: path
          required: true
          description: Version sequence number to update.
          schema:
            type: integer
      x-edge-cases:
        - "Fails with 422 Unprocessable Entity if the version status is not 'draft' (e.g. attempting to update a live/archived template)."
      x-test-coverage:
        - "TestUpdateVersion_Draft: Verifies modification updates template content and returns 200."
        - "TestUpdateVersion_LiveVersionRejected: Verifies that updating a published live template version fails with 422 and 'only draft versions can be modified'."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateVersionRequest"
      responses:
        "200":
          description: Version updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    $ref: "#/components/schemas/PromptVersion"
                required:
                  - version
        "400":
          description: Invalid syntax or missing template
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "template is required"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Version not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "version not found"
        "422":
          description: Only draft versions can be modified
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "only draft versions can be modified"
  /v1/prompts/{id}/versions/{version}/publish:
    post:
      summary: Publish Prompt Version
      description: |
        Publishes a draft version of the prompt template, marking it as 'live'.
        This automatically updates any pre-existing 'live' version to 'archived', maintaining exactly one active live prompt.
      operationId: publishVersion
      tags:
        - Prompts
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: Unique prompt UUID.
          schema:
            type: string
            format: uuid
        - name: version
          in: path
          required: true
          description: Version sequence number to publish.
          schema:
            type: integer
      x-edge-cases:
        - "Switches previous live versions to archived status automatically."
        - "Returns 422 if the specified version is already 'live'."
      x-test-coverage:
        - "TestPublishVersion: Verifies that status changes to 'live' and published_at gets populated."
        - "TestPublishVersion_ArchivesPreviousLive: Verifies that publishing v2 transitions v1 status to archived."
        - "TestPublishVersion_AlreadyLive: Verifies republishing a live template returns 422."
      responses:
        "200":
          description: Version published successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  version:
                    $ref: "#/components/schemas/PromptVersion"
                required:
                  - version
        "400":
          description: Bad parameters
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "invalid version number"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "unauthorized"
        "404":
          description: Version not found
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "version not found"
        "422":
          description: Publishing error
          content:
            application/json:
              schema:
                $ref: "./auth.yaml#/components/schemas/APIError"
              example:
                error: "only draft versions can be modified"
...
components:
  schemas:
...
    PromptVersion:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique version identifier.
        prompt_id:
          type: string
          format: uuid
          description: Reference to parent Prompt ID.
        version:
          type: integer
          description: Incrementing sequence number of the version.
        template:
          type: string
          description: The template syntax code with Go template parameters.
          example: "Hello {{.name}}!"
        status:
          type: string
          enum: [draft, live, archived]
          description: Active lifecycle status of this template version.
        created_at:
          type: string
          format: date-time
        published_at:
          type: string
          format: date-time
          nullable: true
          description: Timestamp when status was set to live. Null if draft.
      required:
        - id
        - prompt_id
        - version
        - template
        - status
        - created_at
        - published_at
...
    CreateVersionRequest:
      type: object
      properties:
        template:
          type: string
          example: "Hello, {{.name}}!"
      required:
        - template
    UpdateVersionRequest:
      type: object
      properties:
        template:
          type: string
          example: "Hello, {{.name}}! Count is {{.count}}."
      required:
        - template
...